requirements.txt
numpy==1.24.3
matplotlib==3.7.1
google-generativeai==0.3.1
python-dotenv==1.0.0 